norm 100% + buf leaks free() >> to pc
